refactor : 네이밍 변경, 바우처 value 자료형 변경 printInputErrorMessage 에서 printErrorMessage 로 변경 => 사용자 입력 에러 뿐만 아니라 다른 에러 메시지 출력도 하기 위해서 자료형 => int 에서 long으로 변경

+++ src/main/java/org/programmers/kdt/weekly/voucher/model/FixedAmountVoucher.java
@@ -6,7 +6,7 @@ import java.util.UUID;
 public class FixedAmountVoucher implements Voucher {
 
     private final UUID voucherId;
-    private int amount;
+    private long amount;
     private final LocalDateTime createdAt;
     private static final VoucherType voucherType = VoucherType.FIXED_AMOUNT_VOUCHER;
 
@@ -32,7 +32,7 @@ public class FixedAmountVoucher implements Voucher {
     }
 
     @Override
-    public int getValue() {
+    public long getValue() {
         return amount;
     }
